Move all esmvaltool specific stuff to own package
This make base package much cleaner
