--- src/ewatercycle/base/recipe_generic_distributed.yml
@@ -1,50 +1,0 @@
-documentation:
-  title: Generate forcing for a generic distributed hydrological model
-  description: |
-    ESMValTool recipe to generate default forcing data for a distributed model in eWaterCycle
-
-  authors:
-    - unmaintained
-
-  maintainer:
-    - unmaintained
-
-  projects:
-    - ewatercycle
-
-  references:
-    - acknow_project
-
-preprocessors:
-  preprocessor: &general
-    extract_shape:
-      shapefile: Meuse_HYPE.shp
-  temperature:
-    <<: *general
-    convert_units:
-      units: degC
-  water:
-    <<: *general
-    convert_units:
-      units: kg m-2 d-1
-
-datasets:
-  - {dataset: ERA5, project: OBS6, tier: 3, type: reanaly, version: 1}
-
-diagnostics:
-  diagnostic:
-    description: HYPE input preprocessor for daily data
-    variables:
-      tas: &var
-        mip: day
-        start_year: 1990
-        end_year: 2001
-        preprocessor: temperature
-      tasmin: *var
-      tasmax: *var
-      pr:
-        <<: *var
-        preprocessor: water
-    scripts: null
-# Each var is written to a netcdf file
-# for example preproc/diagnostic/pr/OBS6_ERA5_reanaly_1_day_pr_1990-1990.nc
