+++ src/ewatercycle/esmvaltool/datasets.py
@@ -1,0 +1,23 @@
+DATASETS = {
+    "ERA5": {
+        "dataset": "ERA5",
+        "project": "OBS6",
+        "tier": 3,
+        "type": "reanaly",
+        "version": 1,
+        "mip": "day",
+    },
+    "ERA-Interim": {
+        "dataset": "ERA-Interim",
+        "project": "OBS6",
+        "tier": 3,
+        "type": "reanaly",
+        "version": 1,
+        "mip": "day",
+    },
+}
+"""Dictionary of predefined forcing datasets.
+
+Where key is the name of the dataset and
+value is an `ESMValTool dataset section <https://docs.esmvaltool.org/projects/ESMValCore/en/latest/recipe/overview.html#datasets>`_.
+"""
