+++ src/ewatercycle/plugins/hype/forcing.py
@@ -6,7 +6,9 @@ import pandas as pd
 import xarray as xr
 from esmvalcore.experimental import get_recipe
 
-from ewatercycle.base.forcing import DATASETS, DefaultForcing, run_esmvaltool_recipe
+from ewatercycle.base.forcing import DefaultForcing
+from ewatercycle.esmvaltool.datasets import DATASETS
+from ewatercycle.esmvaltool.run import run_esmvaltool_recipe
 from ewatercycle.util import get_time, to_absolute_path
